Refined logic for FC, BC, and TT functionality. See KnowledgeBase.Query and new classes for FC, BC, and TT. Also fixed some class responsibilities.
Program class has responsibility for handling console output.
FileHandler class has responsibility for reading the input file, and feeding/querying the knowledge base depending on file input.
KnowledgeBase class has the responsibility to store terms and statements, and instantiate the required methods for FC/BC/TT.
FC/BC/TT classes take the statements and terms from knowledgebase and does operations. Query returns whether or not the supplied query is entailable. There is a seperate function for getting the entailable terms or whatever the case may be.